+++ sources/Renderer/Direct3D11/Buffer/D3D11StreamOutputBuffer.cpp
@@ -15,7 +15,7 @@ namespace LLGL
 D3D11StreamOutputBuffer::D3D11StreamOutputBuffer(ID3D11Device* device, const BufferDescriptor& desc, const void* initialData) :
     D3D11Buffer { BufferType::StreamOutput }
 {
-    CreateResource(device, CD3D11_BUFFER_DESC(desc.size, D3D11_BIND_STREAM_OUTPUT), initialData, desc.flags);
+    CreateResource(device, CD3D11_BUFFER_DESC(static_cast<UINT>(desc.size), D3D11_BIND_STREAM_OUTPUT), initialData, desc.flags);
 }
